requirements.txt
a2wsgi==1.10.1
alembic==1.13.1
anyio==4.2.0
asgiref==3.7.2
attrs==23.2.0
bcrypt==4.1.2
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
connexion==3.0.6
Flask==3.0.2
google-api-core==2.17.1
google-api-python-client==2.120.0
google-auth==2.28.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
googlemaps==4.10.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.3
httplib2==0.22.0
httptools==0.6.1
httpx==0.26.0
idna==3.6
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
Mako==1.3.2
MarkupSafe==2.1.5
networkx==3.3
oauthlib==3.2.2
polyline==2.0.2
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyJWT==2.8.0
pyodbc==5.1.0
pyparsing==3.1.1
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
redis==5.0.7
referencing==0.33.0
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.18.0
rsa==4.9
sniffio==1.3.0
SQLAlchemy==2.0.27
starlette==0.37.1
swagger_ui_bundle==1.1.0
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.27.1
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.1
